Moved render functions for Episodes into renderUtils, and moved the data objects into data.jsx.
