+++ src/components/Episodes/index.js
@@ -1,1 +1,5 @@
+import data from "./data.js";
+import RENDER from "./renderUtils.jsx";
+
+export { data, RENDER };
 export { default } from "./Episodes.jsx";
